refactor(build-optimizer): add sets selects component
